Feat: Hide/Show cursor on window focus events
In order to allow the cursor to disappear when the terminal where we are
running loses focus, this commit updates bubbletea to v1.1.1. It also
handles regaining focus and starts the cursor blinking back up. This is
useful when we are run in a terminal multiplexer like tmux and the
terminal can lose/gain focus often.